+++ src/js/worker/util/changes.js
@@ -214,6 +214,11 @@ const all = [
         msg:
             'Lots of minor UI improvements, especially on mobile. And <a href="/options">a dark theme is available in the options</a>! <a href="https://basketball-gm.com/blog/2018/10/lots-of-small-ui-improvements/">Read about it here.</a> Also since sometimes I forget to update the in-game notifications, check out <a href="https://basketball-gm.com/blog/">the BBGM blog</a> for other news.',
     },
+    {
+        date: "2018-10-15",
+        msg:
+            'You can mark players and picks as "untouchable" when making a trade, to see if the AI can make a counter offer without them. <a href="https://basketball-gm.com/blog/2018/10/untouchable-players-and-picks-in-trades/">More details.</a>',
+    },
 ];
 
 const check = async (conditions: Conditions) => {